Add Reactor.signal() instead of distinctAndMapToValue() and distinctAndCompactMapToValue()
